requirements.txt
fastapi
uvicorn[standard]
pydantic
faiss-cpu
sentence-transformers
numpy
requests
python-dotenv
azure-ai-formrecognizer
python-multipart
